karaboFAI/algorithms/data_structures.py
"""
Offline and online data analysis and visualization tool for azimuthal
integration of different data acquired with various detectors at
European XFEL.

Data structures.

Author: Jun Zhu <jun.zhu@xfel.eu>
Copyright (C) European X-Ray Free-Electron Laser Facility GmbH.
All rights reserved.
"""


class Stack:
    """An FIFO stack."""
    def __init__(self):
        self.__items = []

    def push(self, item):
        self.__items.append(item)

    def pop(self):
        return self.__items.pop(-1)

    def empty(self):
        return not self.__items

    def __len__(self):
        return len(self.__items)
